Align the bundle producer with spec 5e54be4

+++ AddOns/Speckle/Sources/AddOn/DataTypes/Material.h
@@ -10,7 +10,9 @@ struct Material
     std::string applicationId = GuidGenerator::NewGuid();
     uint32_t diffuse = 0;
     double opacity = 1.0;
-    double emissive = 0.0;
+    // Packed ARGB emission colour. 0 = no emission; the bundle writes NULL for it,
+    // per the spec's "producers normalize black RGB to NULL" contract.
+    uint32_t emissive = 0;
     double metalness = 0.0;
     double roughness = 1.0;
 };
